Update tests for 'cleaning' functions, and drop test_GFFcleaner.py (tests moved to 'cleaning' unit tests).
