+++ test/clean/test_sgd.py
@@ -8,26 +8,36 @@ from GFFUtils.clean.sgd import *
 
 class TestGroupByID(unittest.TestCase):
 
+    def setUp(self):
+        # List of genes to group
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=CDS:YEL0W01:1;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=CDS:YEL0W02:1;SGD=YEL0W02
+chr1\tTest\tCDS\t32611\t34140\t0\t-\t0\tID=CDS:YEL0W02:2;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=CDS:YEL0W03:1;SGD=YEL0W03
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=CDS:YEL0W04:1;SGD=YEL0W04
+chr2\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=CDS:YEL0W05:1;SGD=YEL0W05
+chr2\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=CDS:YEL0W05:2;SGD=YEL0W05
+chr2\tTest\tCDS\t40406\t40864\t0\t-\t0\tID=CDS:YEL0W05:3;SGD=YEL0W05
+chr2\tTest\tCDS\t41402\t41831\t0\t-\t0\tID=CDS:YEL0W06:1;SGD=YEL0W06
+""")
+
     def test_group_by_id(self):
         """
         GroupByID: returns features grouped by ID
         """
-        gff = GFFFile("test.gff",
-                      fp=StringIO(u"""5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=CDS:YEL065W:1;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2
-5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=CDS:YDR418W:1;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14
-5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=CDS:YDR418W:2;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15
-2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=CDS:YDR418W:3;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593
-11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=CDS:YKR100C:1;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338
-11	Scannell_and Zill 2011	CDS	632510	635710	Anc_8.301	+	0	ID=1690892742067829512;SGD=;Gene=Sbay_11.340;Parent=Sbay_11.340
-"""))
+        gff = GFFFile('test.gff',self.fp)
         groups = GroupByID(gff)
-        self.assertEqual(len(groups),4)
-        expected_group_ids = ((u"CDS:YEL065W:1",),
-                              (u"CDS:YDR418W:1",
-                               u"CDS:YDR418W:2",
-                               u"CDS:YDR418W:3",),
-                              (u"CDS:YKR100C:1",),
-                              (u"1690892742067829512",))
+        self.assertEqual(len(groups),6)
+        expected_group_ids = ((u"CDS:YEL0W01:1",),
+                              (u"CDS:YEL0W02:1",
+                               u"CDS:YEL0W02:2",),
+                              (u"CDS:YEL0W03:1",),
+                              (u"CDS:YEL0W04:1",),
+                              (u"CDS:YEL0W05:1",
+                               u"CDS:YEL0W05:2",
+                               u"CDS:YEL0W05:3",),
+                              (u"CDS:YEL0W06:1",))
         for group,expected_group in zip(groups,expected_group_ids):
             for idx,expected_idx in zip(group,expected_group):
                 self.assertEqual(idx['attributes']['ID'],
@@ -35,118 +45,365 @@ class TestGroupByID(unittest.TestCase):
 
 class TestGFFGetDuplicateSGDs(unittest.TestCase):
 
+    def setUp(self):
+        # Make file-like object to read data in
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=YEL0W01;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t32611\t34140\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=YEL0W03;SGD=YEL0W03
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=YEL0W04;SGD=YEL0W04
+chr1\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=YEL0W05;SGD=YEL0W05
+chr1\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=YEL0W05;SGD=YEL0W05
+chr2\tTest\tCDS\t40406\t40864\t0\t+\t0\tID=YEL0W05;SGD=YEL0W05
+chr2\tTest\tCDS\t41402\t41831\t0\t-\t0\tID=YEL0W06;SGD=YEL0W06
+""")
+
     def test_gff_get_duplicate_sgds(self):
         """
         GFFGetDuplicateSGDs: fetch GFF records with duplicated 'SGD' values
         """
-        gff = GFFFile("test.gff",
-                      fp=StringIO(u"""5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=1690892742066571750;ygobhmm=2.2e-19;none=;manual=-1e+100;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=0.0;hcnf=0;aa=0.0;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_8.272;stop=0;introns=0;BLAST=YEL065W;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1884;hmm=-1e+100;Name=Sbay_5.2;ty=1e+100
-5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;ygobhmm=7.9e-55;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14;kaks2=-1e+100;struct=1;ltr=0.33;hsp=8e-33;hcnf=0;aa=8e-33;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=261;hmm=-1e+100;Name=Sbay_5.14;ty=1e+100
-5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=1690892742066567879;ygobhmm=1.5e-108;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=4e-65;hcnf=0;aa=4e-65;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=495;hmm=-1e+100;Name=Sbay_5.15;ty=1e+100
-2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;ygobhmm=3.2e-111;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593;kaks2=-1e+100;struct=2;ltr=0.64;hsp=1e-67;hcnf=0;aa=1e-67;lda=-1e+100;rna=-1e+100;synt=3;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=498;hmm=-1e+100;Name=Sbay_2.593;ty=1e+100
-11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=1690892742067852754;ygobhmm=2.4e-104;none=;manual=-1e+100;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338;kaks2=-1e+100;struct=2;ltr=1.4;hsp=6e-115;hcnf=0;aa=6e-115;lda=-1e+100;rna=-1e+100;synt=4;YGOB=Anc_5.716;stop=0;introns=0;BLAST=YKR100C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1074;hmm=-1e+100;Name=Sbay_11.338;ty=5.7
-"""))
+        gff = GFFFile('test.gff',self.fp)
         duplicates = GFFGetDuplicateSGDs(gff)
-        self.assertEqual(list(duplicates.keys()),
-                         ['YDR418W'])
-        expected_duplicates = (
-            "5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;ygobhmm=7.9e-55;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14;kaks2=-1e+100;struct=1;ltr=0.33;hsp=8e-33;hcnf=0;aa=8e-33;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=261;hmm=-1e+100;Name=Sbay_5.14;ty=1e+100",
-            "5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=1690892742066567879;ygobhmm=1.5e-108;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=4e-65;hcnf=0;aa=4e-65;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=495;hmm=-1e+100;Name=Sbay_5.15;ty=1e+100",
-            "2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;ygobhmm=3.2e-111;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593;kaks2=-1e+100;struct=2;ltr=0.64;hsp=1e-67;hcnf=0;aa=1e-67;lda=-1e+100;rna=-1e+100;synt=3;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=498;hmm=-1e+100;Name=Sbay_2.593;ty=1e+100",)
-        for dup,expected in zip(duplicates['YDR418W'],
-                                expected_duplicates):
-            self.assertEqual(str(dup),expected)
+        # Check duplicates
+        self.assertEqual(len(duplicates.keys()),2,
+                         "wrong number of duplicate SGDs")
+        self.assertTrue('YEL0W02' in duplicates.keys())
+        self.assertEqual(len(duplicates['YEL0W02']),2)
+        self.assertTrue('YEL0W05' in duplicates.keys())
+        self.assertEqual(len(duplicates['YEL0W05']),3)
+
+class TestGFFGetDuplicateSGDs(unittest.TestCase):
+
+    def setUp(self):
+        # Make file-like object to read data in
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=YEL0W01;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t32611\t34140\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=YEL0W03;SGD=YEL0W03
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=YEL0W04;SGD=YEL0W04
+chr1\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=YEL0W05;SGD=YEL0W05
+chr1\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=YEL0W05;SGD=YEL0W05
+chr2\tTest\tCDS\t40406\t40864\t0\t+\t0\tID=YEL0W05;SGD=YEL0W05
+chr2\tTest\tCDS\t41402\t41831\t0\t-\t0\tID=YEL0W06;SGD=YEL0W06
+""")
+    
+    def test_get_duplicate_sgds(self):
+        """Test identifying and returning lines with duplicate SGDs
+        """
+        gff = GFFFile('test.gff',self.fp)
+        duplicates = GFFGetDuplicateSGDs(gff)
+        # Check duplicates
+        self.assertEqual(len(duplicates.keys()),2,
+                         "wrong number of duplicate SGDs")
+        self.assertTrue('YEL0W02' in duplicates.keys())
+        self.assertEqual(len(duplicates['YEL0W02']),2)
+        self.assertTrue('YEL0W05' in duplicates.keys())
+        self.assertEqual(len(duplicates['YEL0W05']),3)
 
 class TestGFFResolveDuplicateSGDs(unittest.TestCase):
 
-    def test_gff_resolve_duplicate_sgds(self):
+    def setUp(self):
+        # Make file-like object with data for duplicate
+        # resolution test
+        #
+        # Possible duplicate cases are:
+        # - unrelated duplicates in same chromosome (YEL0W01)
+        # - unrelated duplicates in different chromosomes (YEL0W2)
+        # - grouped duplicated in same chromosome (YEL0W03)
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=CDS:YEL0W01:1;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=CDS:YEL0W02:1;SGD=YEL0W02
+chr1\tTest\tCDS\t32611\t34140\t0\t-\t0\tID=CDS:YEL0W02:2;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=CDS:YEL0W03:1;SGD=YEL0W03
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=CDS:YEL0W03:2;SGD=YEL0W03
+chr1\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=CDS:YEL0W04:1;SGD=YEL0W04
+chr1\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=CDS:YEL0W01:1;SGD=YEL0W01
+chr2\tTest\tCDS\t40406\t40864\t0\t-\t0\tID=CDS:YEL0W02:1;SGD=YEL0W02
+chr2\tTest\tCDS\t41402\t41831\t0\t+\t0\tID=CDS:YEL0W05:1;SGD=YEL0W05
+""")
+        # Mapping data to resolve all duplicates
+        self.mp_resolve_all = StringIO(
+u"""YEL0W01\tchr1\t39195\t39569\t-
+YEL0W03\tchr1\t34525\t37004\t-
+YEL0W02\tchr2\t40406\t40864\t-
+""")
+        # Mapping data with a mapping gene removed
+        self.mp_missing_mapping_gene = StringIO(
+u"""YEL0W01\tchr1\t39195\t39569\t-
+YEL0W03\tchr1\t34525\t37004\t-
+""")
+        # Mapping data with a mapping gene that doesn't match
+        # on strand for one duplicate
+        self.mp_missing_matching_mapping_gene = StringIO(
+u"""YEL0W01\tchr1\t39195\t39569\t-
+YEL0W03\tchr1\t34525\t37004\t+
+YEL0W02\tchr2\t40406\t40864\t-
+""")
+        # Mapping data with a mapping gene that doesn't overlap
+        self.mp_mapping_gene_no_overlap = StringIO(
+u"""YEL0W03\tchr1\t34525\t37004\t-
+YEL0W01\tchr1\t41402\t41831\t-
+YEL0W02\tchr2\t40406\t40864\t-
+""")
+        # Mapping data with multiple mapping genes with same name
+        self.mp_multiple_mapping_genes = StringIO(
+u"""YEL0W01\tchr1\t28789\t29049\t-
+YEL0W03\tchr1\t34525\t37004\t-
+YEL0W01\tchr1\t39195\t39569\t-
+YEL0W02\tchr2\t40406\t40864\t-
+""")
+    
+    def test_resolve_duplicate_sgds(self):
         """
-        GFFResolveDuplicateSGDs: resolve duplicate SGDs using mapping file
+        GFFResolveDuplicateSGDs: all duplicates can be resolved
         """
-        gff = GFFFile("test.gff",
-                      fp=StringIO(u"""chr5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=1690892742066571750;ygobhmm=2.2e-19;none=;manual=-1e+100;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=0.0;hcnf=0;aa=0.0;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_8.272;stop=0;introns=0;BLAST=YEL065W;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1884;hmm=-1e+100;Name=Sbay_5.2;ty=1e+100
-chr5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;ygobhmm=7.9e-55;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14;kaks2=-1e+100;struct=1;ltr=0.33;hsp=8e-33;hcnf=0;aa=8e-33;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=261;hmm=-1e+100;Name=Sbay_5.14;ty=1e+100
-chr2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;ygobhmm=3.2e-111;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593;kaks2=-1e+100;struct=2;ltr=0.64;hsp=1e-67;hcnf=0;aa=1e-67;lda=-1e+100;rna=-1e+100;synt=3;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=498;hmm=-1e+100;Name=Sbay_2.593;ty=1e+100
-chr11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=1690892742067852754;ygobhmm=2.4e-104;none=;manual=-1e+100;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338;kaks2=-1e+100;struct=2;ltr=1.4;hsp=6e-115;hcnf=0;aa=6e-115;lda=-1e+100;rna=-1e+100;synt=4;YGOB=Anc_5.716;stop=0;introns=0;BLAST=YKR100C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1074;hmm=-1e+100;Name=Sbay_11.338;ty=5.7
-"""))
-        mapping_data = TabFile(fp=StringIO(u"""YEL065W	chr5	271736	353971	+
-YDR418W	chr5	28894	29043	-
-YKR100C	chr5	1050499	1112150	-
-"""),
-                               column_names=('name',
-                                             'chr',
-                                             'start',
-                                             'end',
-                                             'strand'))
+        # Load data
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp_resolve_all,
+                          column_names=('name','chr','start','end','strand'))
+        # Fetch duplicates
         duplicates = GFFGetDuplicateSGDs(gff)
-        results = GFFResolveDuplicateSGDs(gff,
-                                          mapping_data,
-                                          duplicates)
-        self.assertEqual(results['resolved_sgds'],['YDR418W'])
-        self.assertEqual(results['unresolved_sgds'],[])
-        self.assertEqual(results['unresolved_sgds_no_mapping_genes'],[])
-        self.assertEqual(results['unresolved_sgds_no_mapping_genes_after_filter'],[])
-        self.assertEqual(results['unresolved_sgds_no_overlaps'],[])
-        self.assertEqual(results['unresolved_sgds_multiple_matches'],[])
-        self.assertEqual(len(results['discard']),1)
+        self.assertEqual(len(duplicates),3,"wrong number of duplicates returned")
+        # Resolve
+        result = GFFResolveDuplicateSGDs(gff,mapping,duplicates,1000)
+        # Check results of resolution (GFF data should be unchanged)
+        #
+        # Selected for discard
+        discard = result["discard"]
+        self.assertEqual(len(discard),3,"wrong number of duplicates marked for discard")
+        self.assertTrue(discard[0] == gff[0])
+        self.assertTrue(discard[1] == gff[1])
+        self.assertTrue(discard[2] == gff[2])
+        #
+        # Resolved SGDs
+        resolved_sgds = result["resolved_sgds"]
+        self.assertEqual(len(resolved_sgds),3)
+        self.assertTrue('YEL0W01' in resolved_sgds)
+        self.assertTrue('YEL0W02' in resolved_sgds)
+        self.assertTrue('YEL0W03' in resolved_sgds)
+        #
+        # Unresolved SGDs
+        self.assertEqual(len(result["unresolved_sgds"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes_after_filter"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_overlaps"]),0)
+        self.assertEqual(len(result["unresolved_sgds_multiple_matches"]),0)
+    
+    def test_resolve_duplicate_sgds_no_mapping_gene(self):
+        """
+        GFFResolveDuplicateSGDs: missing mapping gene
+        """
+        # Load data
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp_missing_mapping_gene,
+                          column_names=('name','chr','start','end','strand'))
+        # Fetch duplicates
+        duplicates = GFFGetDuplicateSGDs(gff)
+        self.assertEqual(len(duplicates),3,"wrong number of duplicates returned")
+        # Resolve
+        result = GFFResolveDuplicateSGDs(gff,mapping,duplicates,1000)
+        # Check results of resolution (GFF data should be unchanged)
+        #
+        # Selected for discard
+        discard = result["discard"]
+        self.assertEqual(len(discard),1,"wrong number of duplicates marked for discard")
+        self.assertTrue(discard[0] == gff[0])
+        #
+        # Resolved SGDs
+        resolved_sgds = result["resolved_sgds"]
+        self.assertEqual(len(resolved_sgds),2)
+        self.assertTrue('YEL0W01' in resolved_sgds)
+        self.assertTrue('YEL0W03' in resolved_sgds)
+        #
+        # Unresolved SGDs
+        self.assertEqual(len(result["unresolved_sgds"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes_after_filter"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_overlaps"]),0)
+        self.assertEqual(len(result["unresolved_sgds_multiple_matches"]),0)
+        self.assertTrue('YEL0W02' in result["unresolved_sgds_no_mapping_genes"])
+        self.assertTrue('YEL0W02' in result["unresolved_sgds"])
+
+    def test_resolve_duplicate_sgds_no_matching_mapping_gene(self):
+        """
+        GFFResolveDuplicateSGDs: no matching mapping gene
+        """
+        # Load data
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp_missing_matching_mapping_gene,
+                          column_names=('name','chr','start','end','strand'))
+        # Fetch duplicates
+        duplicates = GFFGetDuplicateSGDs(gff)
+        self.assertEqual(len(duplicates),3,"wrong number of duplicates returned")
+        # Resolve
+        result = GFFResolveDuplicateSGDs(gff,mapping,duplicates,1000)
+        # Check results of resolution (GFF data should be unchanged)
+        #
+        # Selected for discard
+        discard = result["discard"]
+        self.assertEqual(len(discard),3,"wrong number of duplicates marked for discard")
+        self.assertTrue(discard[0] == gff[0])
+        self.assertTrue(discard[1] == gff[1])
+        self.assertTrue(discard[2] == gff[2])
+        #
+        # Resolved SGDs
+        resolved_sgds = result["resolved_sgds"]
+        self.assertEqual(len(resolved_sgds),2)
+        self.assertTrue('YEL0W01' in resolved_sgds)
+        self.assertTrue('YEL0W02' in resolved_sgds)
+        #
+        # Unresolved SGDs
+        self.assertEqual(len(result["unresolved_sgds"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes_after_filter"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_overlaps"]),0)
+        self.assertEqual(len(result["unresolved_sgds_multiple_matches"]),0)
+        self.assertTrue('YEL0W03' in result["unresolved_sgds_no_mapping_genes_after_filter"])
+        self.assertTrue('YEL0W03' in result["unresolved_sgds"])
+
+    def test_resolve_duplicate_sgds_mapping_gene_with_no_overlap(self):
+        """
+        GFFResolveDuplicateSGDs: mapping gene with no overlap
+        """
+        # Load data
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp_mapping_gene_no_overlap,
+                          column_names=('name','chr','start','end','strand'))
+        # Fetch duplicates
+        duplicates = GFFGetDuplicateSGDs(gff)
+        self.assertEqual(len(duplicates),3,"wrong number of duplicates returned")
+        # Resolve
+        result = GFFResolveDuplicateSGDs(gff,mapping,duplicates,1000)
+        # Check results of resolution (GFF data should be unchanged)
+        #
+        # Selected for discard
+        discard = result["discard"]
+        self.assertEqual(len(discard),2,"wrong number of duplicates marked for discard")
+        self.assertTrue(discard[0] == gff[1])
+        self.assertTrue(discard[1] == gff[2])
+        #
+        # Resolved SGDs
+        resolved_sgds = result["resolved_sgds"]
+        self.assertEqual(len(resolved_sgds),2)
+        self.assertTrue('YEL0W02' in resolved_sgds)
+        self.assertTrue('YEL0W03' in resolved_sgds)
+        #
+        # Unresolved SGDs
+        self.assertEqual(len(result["unresolved_sgds"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes_after_filter"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_overlaps"]),1)
+        self.assertEqual(len(result["unresolved_sgds_multiple_matches"]),0)
+        self.assertTrue('YEL0W01' in result["unresolved_sgds_no_overlaps"])
+        self.assertTrue('YEL0W01' in result["unresolved_sgds"])
+
+    def test_resolve_duplicate_sgds_multiple_mapping_genes_with_same_name(self):
+        """
+        GFFResolveDuplicateSGDs: multiple mapping genes with the same name
+        """
+        # Load data
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp_multiple_mapping_genes,
+                          column_names=('name','chr','start','end','strand'))
+        # Fetch duplicates
+        duplicates = GFFGetDuplicateSGDs(gff)
+        self.assertEqual(len(duplicates),3,"wrong number of duplicates returned")
+        # Resolve
+        result = GFFResolveDuplicateSGDs(gff,mapping,duplicates,1000)
+        # Check results of resolution (GFF data should be unchanged)
+        #
+        # Selected for discard
+        discard = result["discard"]
+        self.assertEqual(len(discard),2,"wrong number of duplicates marked for discard")
+        self.assertTrue(discard[0] == gff[1])
+        self.assertTrue(discard[1] == gff[2])
+        #
+        # Resolved SGDs
+        resolved_sgds = result["resolved_sgds"]
+        self.assertEqual(len(resolved_sgds),2)
+        self.assertTrue('YEL0W02' in resolved_sgds)
+        self.assertTrue('YEL0W03' in resolved_sgds)
+        #
+        # Unresolved SGDs
+        self.assertEqual(len(result["unresolved_sgds"]),1)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_mapping_genes_after_filter"]),0)
+        self.assertEqual(len(result["unresolved_sgds_no_overlaps"]),0)
+        self.assertEqual(len(result["unresolved_sgds_multiple_matches"]),1)
+        self.assertTrue('YEL0W01' in result["unresolved_sgds_multiple_matches"])
+        self.assertTrue('YEL0W01' in result["unresolved_sgds"])
 
 class TestGFFGroupSGDs(unittest.TestCase):
+
+    def setUp(self):
+        # Make file-like object to read data in
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=YEL0W01;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t32611\t34140\t0\t-\t0\tID=YEL0W02;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=YEL0W03;SGD=YEL0W03
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=YEL0W03;SGD=YEL0W03
+chr1\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=YEL0W04;SGD=YEL0W04
+chr1\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=YEL0W03;SGD=YEL0W03
+chr1\tTest\tCDS\t40406\t40864\t0\t-\t0\tID=YEL0W01;SGD=YEL0W01
+""")
+        # The expected ID assignments for the input above
+        self.expected_ids = ("CDS:YEL0W01:1",
+                             "CDS:YEL0W02:1",
+                             "CDS:YEL0W02:2",
+                             "CDS:YEL0W03:1",
+                             "CDS:YEL0W03:2",
+                             "CDS:YEL0W04:1",
+                             "CDS:YEL0W03:3",
+                             "CDS:YEL0W01:1")
 
     def test_gff_group_sgds(self):
         """
         GFFGroupSGDs: relabel ID with SGD information
         """
-        gff = GFFFile("test.gff",
-                      fp=StringIO(u"""5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=1690892742066571750;ygobhmm=2.2e-19;none=;manual=-1e+100;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=0.0;hcnf=0;aa=0.0;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_8.272;stop=0;introns=0;BLAST=YEL065W;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1884;hmm=-1e+100;Name=Sbay_5.2;ty=1e+100
-5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;ygobhmm=7.9e-55;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14;kaks2=-1e+100;struct=1;ltr=0.33;hsp=8e-33;hcnf=0;aa=8e-33;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=261;hmm=-1e+100;Name=Sbay_5.14;ty=1e+100
-5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=1690892742066567879;ygobhmm=1.5e-108;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15;kaks2=-1e+100;struct=2;ltr=1e+100;hsp=4e-65;hcnf=0;aa=4e-65;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=495;hmm=-1e+100;Name=Sbay_5.15;ty=1e+100
-2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;ygobhmm=3.2e-111;none=;manual=-1e+100;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593;kaks2=-1e+100;struct=2;ltr=0.64;hsp=1e-67;hcnf=0;aa=1e-67;lda=-1e+100;rna=-1e+100;synt=3;YGOB=Anc_5.522;stop=0;introns=0;BLAST=YEL054C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=498;hmm=-1e+100;Name=Sbay_2.593;ty=1e+100
-11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=1690892742067852754;ygobhmm=2.4e-104;none=;manual=-1e+100;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338;kaks2=-1e+100;struct=2;ltr=1.4;hsp=6e-115;hcnf=0;aa=6e-115;lda=-1e+100;rna=-1e+100;synt=4;YGOB=Anc_5.716;stop=0;introns=0;BLAST=YKR100C;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=1074;hmm=-1e+100;Name=Sbay_11.338;ty=5.7
-11	Scannell_and Zill 2011	CDS	632510	635710	Anc_8.301	+	0	ID=1690892742067829512;ygobhmm=9e-33;none=;manual=-1e+100;SGD=;Gene=Sbay_11.340;Parent=Sbay_11.340;kaks2=-1e+100;struct=-1;ltr=1.1;hsp=3e-86;hcnf=0;aa=3e-86;lda=-1e+100;rna=-1e+100;synt=0;YGOB=Anc_8.301;stop=0;introns=0;BLAST=YKR102W;ncbi=1e+100;kaks=1e+100;nnnn=-1e+100;length=3201;hmm=-1e+100;Name=Sbay_11.340;ty=1e+100
-"""))
+        gff = GFFFile('test.gff',self.fp)
         GFFGroupSGDs(gff)
-        expected_ids = ("CDS:YEL065W:1",
-                        "CDS:YDR418W:1",
-                        "CDS:YDR418W:2",
-                        "CDS:YDR418W:3",
-                        "CDS:YKR100C:1",
-                        "1690892742067829512")
-        for line,expected_id in zip(gff,expected_ids):
+        for line,expected_id in zip(gff,self.expected_ids):
             self.assertEqual(line['attributes']['ID'],expected_id)
 
 class TestGFFInsertMissingGenes(unittest.TestCase):
+
+    def setUp(self):
+        # Make file-like object for GFF data
+        self.fp = StringIO(
+u"""chr1\tTest\tCDS\t28789\t29049\t0\t-\t0\tID=CDS:YEL0W01:1;SGD=YEL0W01
+chr1\tTest\tCDS\t29963\t32155\t0\t-\t0\tID=CDS:YEL0W02:1;SGD=YEL0W02
+chr1\tTest\tCDS\t34525\t35262\t0\t-\t0\tID=CDS:YEL0W04:1;SGD=YEL0W04
+chr1\tTest\tCDS\t35823\t37004\t0\t-\t0\tID=CDS:YEL0W04:2;SGD=YEL0W04
+chr2\tTest\tCDS\t38050\t38120\t0\t-\t0\tID=CDS:YEL0W05:1;SGD=YEL0W05
+chr2\tTest\tCDS\t39195\t39569\t0\t-\t0\tID=CDS:YEL0W06:1;SGD=YEL0W06
+chr2\tTest\tCDS\t40406\t40864\t0\t-\t0\tID=CDS:YEL0W06:2;SGD=YEL0W06
+""")
+        # Make a file-like object for mapping data
+        self.mp = StringIO(
+u"""YEL0W03\tchr1\t32611\t34140\t-
+YEL0W06\tchr2\t49195\t49569\t-
+""")
 
     def test_gff_insert_missing_genes(self):
         """
         GFFInsertMissingGenes: insert gene from mapping data file
         """
-        gff = GFFFile("test.gff",
-                      fp=StringIO(u"""chr5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=1690892742066571750;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2;
-chr5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14
-chr5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=1690892742066567879;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15
-chr2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593
-chr11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=1690892742067852754;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338
-chr11	Scannell_and Zill 2011	CDS	632510	635710	Anc_8.301	+	0	ID=1690892742067829512;SGD=;Gene=Sbay_11.340;Parent=Sbay_11.340
-"""))
-        mapping_data = TabFile(fp=StringIO(u"""YEL065W	chr5	271736	353971	+
-YDR418W	chr5	28894	29043	-
-YKR100C	chr5	1050499	1112150	-
-YIL151C	chr5	1887446	1900607	+
-"""),
-                               column_names=('name',
-                                             'chr',
-                                             'start',
-                                             'end',
-                                             'strand'))
-        GFFInsertMissingGenes(gff,mapping_data)
-        expected_gff = GFFFile("expected.gff",
-                               fp=StringIO(u"""chr5	Scannell_and Zill 2011	CDS	3340	5223	Anc_8.272	+	0	ID=1690892742066571750;SGD=YEL065W;Gene=Sbay_5.2;Parent=Sbay_5.2;
-chr5	Scannell_and Zill 2011	CDS	28789	29049	Anc_5.522	-	0	ID=1690892742066567884;SGD=YDR418W;Gene=Sbay_5.14;Parent=Sbay_5.14
-chr5	Scannell_and Zill 2011	CDS	28789	29313	Anc_5.522	-	0	ID=1690892742066567879;SGD=YDR418W;Gene=Sbay_5.15;Parent=Sbay_5.15
-chr5	GFFcleaner	CDS	1887446	1900607	0	+	0	ID=CDS:YIL151C:1;SGD=YIL151C;Gene=YIL151C;Parent=YIL151C
-chr2	Scannell_and Zill 2011	CDS	1053165	1053662	Anc_5.522	+	0	ID=1690892742067239098;SGD=YDR418W;Gene=Sbay_2.593;Parent=Sbay_2.593
-chr11	Scannell_and Zill 2011	CDS	623811	624884	Anc_5.716	-	0	ID=1690892742067852754;SGD=YKR100C;Gene=Sbay_11.338;Parent=Sbay_11.338
-chr11	Scannell_and Zill 2011	CDS	632510	635710	Anc_8.301	+	0	ID=1690892742067829512;SGD=;Gene=Sbay_11.340;Parent=Sbay_11.340
-"""))
-        for line,expected in zip(gff,expected_gff):
-            self.assertEqual(str(line),str(expected))
+        gff = GFFFile('test.gff',self.fp)
+        mapping = TabFile('map.txt',self.mp,
+                          column_names=('name','chr','start','end','strand'))
+        # Insert missing genes
+        GFFInsertMissingGenes(gff,mapping)
+        # Check: YEL0W03 should be inserted at index 2
+        i = 2
+        idx = gff[i]['attributes']['ID']
+        self.assertEqual(idx,'CDS:YEL0W03:1',"incorrect ID at position %d" % i)
+        # Check: YEL0W06 should not have been inserted
+        for i in range(len(gff)):
+            idx = gff[i]['attributes']['ID']
+            self.assertNotEqual(idx,'CDS:YEL0W06:3',"wrong ID at position %d" %i)
+        # Check: no leading ';' on the string representation
+        self.assertNotEqual(str(gff[i]['attributes'])[0],';',"Erroneous leading semicolon: %s"
+                            % str(gff[i]['attributes']))
